check if file exist before delete

+++ possibilities/pump.js
@@ -73,11 +73,12 @@ mysql_conn.query('INSERT INTO files (pair, dateFrom, dateTo, config_js, strategy
     mysql_conn.end();
     if (err) throw err;
     files_id = result.insertId;
-
-    fs.unlink(APP_DIR+'mysql_tasks.sql', (err) => {
-        if (err) throw err;
-        strategy(combinationTotal);
-    });
+    if (fs.existsSync(APP_DIR+'mysql_tasks.sql')) {
+      fs.unlink(APP_DIR+'mysql_tasks.sql', (err) => {
+          if (err) throw err;
+          strategy(combinationTotal);
+      });
+    }
     
 });
 console.log('--------------------------------------------');
